fix: use per-run UIDs for node runner
