+++ internal/core/runner/nodejs/nodejs_test.go
@@ -1,8 +1,12 @@
 package nodejs
 
 import (
+	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/langgenius/dify-sandbox/internal/core/runner/types"
+	"github.com/langgenius/dify-sandbox/internal/static"
 )
 
 func TestBuildBootstrapInjectsPreloadButNotUserCode(t *testing.T) {
@@ -18,5 +22,21 @@ func TestBuildBootstrapInjectsPreloadButNotUserCode(t *testing.T) {
 
 	if strings.Contains(bootstrap, "console.log('user code')") {
 		t.Fatal("bootstrap unexpectedly contains user code")
+	}
+}
+
+func TestBuildCommandArgsUsesProvidedUID(t *testing.T) {
+	args := buildCommandArgs("/tmp/test.js", 10042, &types.RunnerOptions{})
+
+	if args[0] != "/tmp/test.js" {
+		t.Fatalf("expected script path first, got %q", args[0])
+	}
+
+	if args[1] != "10042" {
+		t.Fatalf("expected provided uid, got %q", args[1])
+	}
+
+	if args[1] == strconv.Itoa(static.SANDBOX_USER_UID) {
+		t.Fatal("expected command args to avoid shared sandbox uid")
 	}
 }
